+++ 예찬님/mysite/contact/templates/contactform/contact.html
@@ -1,6 +1,8 @@
 <!-- <h1>Contact Us</h1>
 <form method="post">
   {% csrf_token %}
+  Please write your name, email address and a message below if you have any questions.
+  One of our staff members will be happy to contact you directly and answer your questions as soon as possible.
   <div>
     <label for="name">Name: </label><br>
     <input id="name" name="name" type="text" placeholder="Enter your name">
