+++ 예찬님/mysite/contact/templates/contactform/contact_complete.html
@@ -1,0 +1,3 @@
+<h3>Contact</h3>
+
+Mail sended
