+++ 예찬님/mysite/mysite/forms.py
@@ -1,6 +1,1 @@
-from django import forms
 
-class ContactForm(forms.Form):
-    from_email = forms.EmailField(required=True)
-    subject = forms.CharField(required=True)
-    message = forms.CharField(widget=forms.Textarea)
