+++ 예찬님/mysite/mysite/settings.py
@@ -37,6 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'contact'
 
 ]
 
@@ -128,8 +129,7 @@ EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
 EMAIL_USE_TLS = True
 EMAIL_HOST = 'smtp.gmail.com'
 EMAIL_PORT = 587
-EMAIL_HOST_USER = 'lee_yechan@hufs.ac.kr' # ex) bum752@gmail.com
-EMAIL_HOST_PASSWORD = '980329' # ex) P@ssw0rd
-SERVER_EMAIL = 'lee_yechan@hufs.ac.kr' # ex) bum752@gmail.com
-
-DEFAULT_FROM_EMAIL  = 'lee_yechan@hufs.ac.kr'
+EMAIL_HOST_USER = 'lee_yechan@hufs.ac.kr'
+EMAIL_HOST_PASSWORD = '980329'
+SERVER_EMAIL = 'example@email.com'
+DEFAULT_FROM_EMAIL  = 'example@email.com'
