+++ 예찬님/mysite/mysite/urls.py
@@ -17,8 +17,12 @@ from django import views
 from django.contrib import admin
 from django.urls import include, path
 
+from contact import views
+
+app_name = "mysite"
+
 urlpatterns = [
     path('polls/', include('polls.urls')),
     path('admin/', admin.site.urls),
-    path("contact/", views.contact, name="contact"),
+    path('contact/', include('contact.urls'), name="contact"),
 ]
